Add GLM library & fix stb image library
We use GLM to do matrix and other mathematical operations because it is
specifically targeted to OpenGL. We could also go for Eigen, but it as
this is a rendering application GLM seems the better choice at the
moment. Also the import optimizer removed the necessary definitions of
the stb image library - it has been included back.