Change interface to C-Style library
To ease integration into other frameworks using different languages a
C-Style interface has been adopted. It just uses primitive types and to
control state and flow of the renderer. This should make it easier to
write wrappers for other languages like Python or C# as there is no need
to serialise custom types and classes.