Connect PyQt5 application to speckle
A simple PyQt5 application has been created, to showcase the connection
of speckle to a desktop applications handling the general GUI, event
loop and other necesities outside of rendering and displaying the point
cloud. This seam has been created early to seperate the GUI from the
rendering engine to stay flexible in terms of the GUI implementation.